class name made more generalised
